+++ packages/gamut/src/Icon/icons/MenuIcon.tsx
@@ -2,7 +2,7 @@ import React, { SVGProps } from 'react';
 
 import { defaultIconProps } from '../defaultIconProps';
 
-export const MenuIcon: React.FC<SVGProps<SVGSVGElement>> = props => {
+export const MenuIcon: React.FC<SVGProps<SVGSVGElement>> = (props) => {
   return (
     <svg {...props}>
       <g fillRule="nonzero">