+++ packages/gamut/src/Icon/icons/PHPIcon.tsx
@@ -2,7 +2,7 @@ import React, { SVGProps } from 'react';
 
 import { defaultIconProps } from '../defaultIconProps';
 
-export const PHPIcon: React.FC<SVGProps<SVGSVGElement>> = props => {
+export const PHPIcon: React.FC<SVGProps<SVGSVGElement>> = (props) => {
   return (
     <svg {...props}>
       <title>PHP Icon</title>