+++ packages/styleguide/stories/Organisms/GridForm.stories.tsx
@@ -75,7 +75,7 @@ export const gridForm = decoratedStory(() => (
           type: 'file',
           validation: {
             required: true,
-            validate: files => {
+            validate: (files) => {
               const { type } = files.item(0);
               const allowedTypes = [
                 'application/pdf',
@@ -131,7 +131,7 @@ export const gridForm = decoratedStory(() => (
           },
         },
       ]}
-      onSubmit={async values => {
+      onSubmit={async (values) => {
         action('Form Submitted')(values);
       }}
       submit={{
@@ -160,7 +160,7 @@ export const gridFormWithSubmitButtonPosition = decoratedStory(() => (
               type: 'text',
             },
           ]}
-          onSubmit={async values => {
+          onSubmit={async (values) => {
             action('Form Submitted')(values);
           }}
           submit={{
@@ -180,7 +180,7 @@ export const gridFormWithSubmitButtonPosition = decoratedStory(() => (
               type: 'text',
             },
           ]}
-          onSubmit={async values => {
+          onSubmit={async (values) => {
             action('Form Submitted')(values);
           }}
           submit={{
@@ -200,7 +200,7 @@ export const gridFormWithSubmitButtonPosition = decoratedStory(() => (
               type: 'text',
             },
           ]}
-          onSubmit={async values => {
+          onSubmit={async (values) => {
             action('Form Submitted')(values);
           }}
           submit={{
@@ -220,7 +220,7 @@ export const gridFormWithSubmitButtonPosition = decoratedStory(() => (
               type: 'text',
             },
           ]}
-          onSubmit={async values => {
+          onSubmit={async (values) => {
             action('Form Submitted')(values);
           }}
           submit={{
@@ -250,7 +250,7 @@ export const gridFormWithSubmitButtonColor = decoratedStory(() => (
           type: 'text',
         },
       ]}
-      onSubmit={async values => {
+      onSubmit={async (values) => {
         action('Form Submitted')(values);
       }}
       submit={{
@@ -267,7 +267,7 @@ export const gridFormWithSubmitButtonColor = decoratedStory(() => (
           type: 'text',
         },
       ]}
-      onSubmit={async values => {
+      onSubmit={async (values) => {
         action('Form Submitted')(values);
       }}
       submit={{
@@ -302,7 +302,7 @@ export const gridFormWithInlineSubmitButton = decoratedStory(() => (
           type: 'checkbox',
         },
       ]}
-      onSubmit={async values => {
+      onSubmit={async (values) => {
         action('Form Submitted')(values);
       }}
       submit={{
@@ -324,7 +324,7 @@ export const gridFormWithInlineSubmitButton = decoratedStory(() => (
           type: 'text',
         },
       ]}
-      onSubmit={async values => {
+      onSubmit={async (values) => {
         action('Form Submitted')(values);
       }}
       submit={{
@@ -360,7 +360,7 @@ export const gridFormWithCustomInput = decoratedStory(() => (
                 error={!!error}
                 id="custom-text-input"
                 type="text"
-                onChange={event => setValue(event.target.value)}
+                onChange={(event) => setValue(event.target.value)}
               />
               <span aria-label="Dancing person" role="img">
                 🕺
@@ -380,7 +380,7 @@ export const gridFormWithCustomInput = decoratedStory(() => (
           type: 'custom',
         },
       ]}
-      onSubmit={async values => {
+      onSubmit={async (values) => {
         action('Form Submitted')(values);
       }}
       submit={{
@@ -418,7 +418,7 @@ export const gridFormWithPlaceholderText = decoratedStory(() => (
           type: 'email',
         },
       ]}
-      onSubmit={async values => {
+      onSubmit={async (values) => {
         action('Form Submitted')(values);
       }}
       submit={{
@@ -450,7 +450,7 @@ export const gridFormWithFieldWithOnUpdate = decoratedStory(() => {
             onUpdate: setText,
           },
         ]}
-        onSubmit={async values => {
+        onSubmit={async (values) => {
           action('Form Submitted')(values);
         }}
         submit={{