+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>7.4.6 (2020-05-05)</small>
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
 ## <small>7.4.5 (2020-05-04)</small>
 
 * chore: Update prettier & Run prettier on packages (#773) ([2f012a1](https://github.com/RyzacInc/client-modules/commit/2f012a1)), closes [#773](https://github.com/RyzacInc/client-modules/issues/773)